refactor(chart): blit fast-path 코드를 chart.blit.js 로 추출
